requirements.txt
requests
pillow
pytesseract
selenium
webdriver-manager
pandas